update: created the application architecture
