--- src/App.js
@@ -1,10 +1,0 @@
-
-function App() {
-  return (
-    <div className="App">
-    
-    </div>
-  );
-}
-
-export default App;
