+++ src/app/components/Header/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Header: React.FunctionComponent = () => {
+    return (
+        <React.Fragment>
+            <h1>Header</h1>
+        </React.Fragment>
+    )
+}
+
+export default Header;
